folder organize + logic improve
